Fix doxm multicast GET request responses
Previously, a random short delay was introduced to responses to
multicast GET doxm requests. During the delay the data was stored
in a single global buffer. However, for simulators with multiple
devices a single buffer is insufficient, a separate buffer is
required for each device.

+++ api/oc_uuid.c
@@ -81,7 +81,7 @@ oc_str_to_uuid(const char *str, oc_uuid_t *uuid)
 }
 
 void
-oc_uuid_to_str(const oc_uuid_t *uuid, char *buffer, int buflen)
+oc_uuid_to_str(const oc_uuid_t *uuid, char *buffer, size_t buflen)
 {
   int i, j = 0;
   if (buflen < OC_UUID_LEN || !uuid)
